Pull request: AGDNS-2848-upd-dnsproxy-doh
Squashed commit of the following:

commit e06d9abfe4de16507d8a87e98b376a241729c010
Merge: 566afb9bf 806855dc1
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Thu Mar 12 08:35:39 2026 +0700

    Merge remote-tracking branch 'origin/master' into AGDNS-2848-upd-dnsproxy-doh

commit 566afb9bf03d282fbe4f150e2db8664c6a0de23b
Merge: 39688da82 354ce04a5
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Wed Mar 11 17:58:44 2026 +0700

    Merge remote-tracking branch 'origin/master' into AGDNS-2848-upd-dnsproxy-doh

commit 39688da82e6f11e1c5e81fd6c0e0f1354bba0970
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Wed Mar 11 08:17:37 2026 +0700

    all: upd dnsproxy

commit 168847c56ceabbf1afdff3ffe0a90102fb53643b
Merge: 7463c0733 a15c72493
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Wed Mar 11 08:09:55 2026 +0700

    Merge remote-tracking branch 'origin/master' into AGDNS-2848-upd-dnsproxy-doh

commit 7463c07339bd6c96ab3d8dd922ee6233aec50a88
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Tue Mar 10 16:32:02 2026 +0700

    all: upd dnsproxy

commit c605ec946588cc2271d2ba83e71986917798c752
Merge: 07da0bbe5 ad9cb3e8d
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Tue Mar 10 08:56:14 2026 +0700

    Merge remote-tracking branch 'origin/master' into AGDNS-2848-upd-dnsproxy-doh

commit 07da0bbe57b3a2836c870d522363f9233ee5dc75
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Mon Mar 9 12:36:11 2026 +0700

    all: imp code

commit bfdf4a4467893ae4b445e8c947e0f60d12675e40
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Mar 6 15:58:20 2026 +0700

    all: upd dnsproxy

commit 097b64b450a2a89b09d8ff066a191fe7122e8de0
Merge: bff075daf 01dd10e49
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Mar 6 12:36:56 2026 +0700

    Merge branch 'master' into AGDNS-2848-upd-dnsproxy-doh
    
    # Conflicts:
    #	go.mod
    #	go.sum
    #	internal/dnsforward/config.go

commit bff075dafbd8dc3482a5cf9d8eb24584713b109b
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Mar 6 12:34:48 2026 +0700

    all: upd dnsproxy

commit 5956d01b0d87a84a6b1155608258dd37949eee4f
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Mar 6 11:59:26 2026 +0700

    all: upd dnsproxy

commit 2e2091e3db9b0309faa476961d136fb1c9bd39e9
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Wed Feb 25 11:31:12 2026 +0700

    all: upd dnsproxy, rm beforerequest

commit 44e5f1207e5201c6527f2f316ceafac2b3593868
Merge: d474280e9 82f2ac68a
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Feb 20 10:38:45 2026 +0700

    Merge remote-tracking branch 'origin/master' into AGDNS-3523-upd-dnsproxy

commit d474280e944ecfab6bfd877089fc7ad79f3ecfed
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Thu Feb 19 20:07:13 2026 +0700

    all: upd dnsproxy

commit b19a7cd851c62fafafed340825e7ad269f5a578a
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Thu Feb 19 13:50:05 2026 +0700

    all: upd dnsproxy

commit 56b67426e0ff804880ac74536f01cdbc4db3605b
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Thu Feb 19 10:06:42 2026 +0700

    all: imp code

commit 716f8186bd5e74152b0677d57a26ca4fc2f991a9
Merge: 96c178bb3 94a3a4fa6
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Thu Feb 19 09:21:49 2026 +0700

    Merge remote-tracking branch 'origin/master' into AGDNS-3523-upd-dnsproxy
    
    # Conflicts:
    #	go.mod
    #	go.sum

commit 96c178bb3d2b6cfcd1cb9377b71c3c23987eab6e
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Tue Feb 17 10:27:27 2026 +0700

    next: dnssvc todos

commit f84834f68f399e6ce841c7c5765f3180acb3c417
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Tue Feb 17 09:11:54 2026 +0700

    all: upd dnsproxy

commit 6ad1a30c087256374882b6e24f5c879121c34834
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Mon Feb 16 13:17:09 2026 +0700

    scripts: fix

commit cc7c6d3181a427f4e79f4af686d3363f5c341001
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Mon Feb 16 13:11:24 2026 +0700

    all: upd golibs

commit cd2662cf6bc52216c9b2ec55ffd34a8ade434170
Merge: d3d2bd6f2 a165cdb68
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Mon Feb 16 11:49:19 2026 +0700

    Merge remote-tracking branch 'refs/remotes/origin/master' into AGDNS-3523-upd-dnsproxy

commit d3d2bd6f27de6b721f73a48e1c4c8e1187d51046
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Feb 6 12:23:43 2026 +0700

    all: todo

commit e16df07e1062e67cc3d085048c4399cf375997bf
Author: Dimitry Kolyshev <dkolyshev@adguard.com>
Date:   Fri Feb 6 12:06:37 2026 +0700

    dnsforward: upd dnsproxy

+++ internal/dnsforward/process.go
@@ -201,7 +201,7 @@ func (s *Server) makeDDRResponse(req *dns.Msg) (resp *dns.Msg) {
 	for _, addr := range s.conf.TLSConf.HTTPSListenAddrs {
 		values := []dns.SVCBKeyValue{
 			&dns.SVCBAlpn{Alpn: []string{"h2"}},
-			&dns.SVCBPort{Port: uint16(addr.Port)},
+			&dns.SVCBPort{Port: addr.Port()},
 			&dns.SVCBDoHPath{Template: "/dns-query{?dns}"},
 		}
